+++ src/features/movies/MovieDetails/Crew/index.js
@@ -3,7 +3,7 @@ import { selectMovieCrew } from "../movieDetailsSlice";
 import { Section } from "../../../../common/Section/styled";
 import { Title } from "../../../../common/Title/styled";
 import { Wrapper } from "../Wrapper";
-import { WrapperLink } from "../../../../common/wrapperLink";
+import { WrapperLink } from "../../../../common/WrapperLink/styled";
 import { PersonTile } from "../../../../common/PersonTile";
 import { apiUrlImage } from "../../../apiData";
 import personError from "../../../../images/personError.jpg";
